Basic functionality for file upload workflow
Closes #1

The module, as of now, exports a single function, fileUploader() which
accepts the following as parameters:

submissionBody -- a string for the body of the submission from the file
being uploaded
submissionFormat -- a string which should specify the format of the
submission body; either 'JSON' or 'XML'
JWT -- the JSON Web Token (a string) to be used when making requests to
the Submissions API
baseSubmissionURL -- a string for the URL to be used (helpful to have it
not default to prod for testing)
callback -- a callback function to be called once the workflow has
finished, which takes an array of Error objects that were thrown as the
first argument and an array of created measurementSet as the second.
The module then follows the business logic outlined in a Confluence
page to validate the submissionBody, query for any existing submissions,
and then using POST /measurement-sets and PUT /measurement-sets
appropriately to save the measurementSets in the submissionBody in the
Submissions API database.

A set of unit tests have been written to test basic functionality as
well.