Fix/qppsf-8721 removing stringify from axios calls, fixing json parsing api error (#143)
* Adding server-test, fixing parsing issues

Since a string is passed into file uploader, adding json.parse to
validation api call and removing the JSON.stringify from the
postMeasurementSet call. Not entirely sure why this is required.

* Removing stringify from axios put

* Removing JSON.stringify from post

Updating stub to properly stringify data as axios would.

* reverting server-test.js changes

* Adding file for testing file-upload.

* removing dead link.

* Removing parse addition to validate call

This was entered incorrectly due to a local environment issue.